Add document method
setFgColor: set text color of the current document.
setBgColor: set background color of the current document.
setLinkColor: set color of links within the document.
setALinkColor: set the color of an active link in the document body.
setVlinkColor: set the color of links that the user has visited in the document.

Delete obsolete method